DB download workflow
This mini-subworkflow will download the databases if the user provide
the dbs flag. It will only download them if the files are not there, but
it doesn't do a proper check, it just checks if the folders exist.
This should be enough provided the pipeline completes those
first few download db steps.